+++ R/EOF.R
@@ -8,6 +8,9 @@
 #' the matrix that will be used in the SVD decomposition (see Details)
 #' @param value.var optional name of the data column (see Details)
 #' @param n which singular values to return (if \code{NULL}, returns all)
+#' @param B number of bootstrap samples used to estimate confidence intervals
+#' @param probs the probabilities of the lower and upper values of estiamted
+#' confidence intervals. If named, it's names will be used as column names.
 #'
 #' @return
 #' A list of 3 named elements containing tidy data.tables of the right and left
@@ -41,6 +44,8 @@
 #' to set n to \code{NULL}. If the irba package is installed, EOF uses
 #' [irlba::irlba] instead of [base::svd] since it's much faster.
 #'
+#' The boostraping procedure follows Fisher et.al. (2016) and returns the
+#' standard deviation of each singular value.
 #'
 #' @examples
 #' # The Antarctic Oscillation is computed from the
@@ -51,7 +56,8 @@
 #' geopotential[, gh.t.w := Anomaly(gh)*sqrt(cos(lat*pi/180)),
 #'       by = .(lon, lat, month(date))]
 #'
-#' aao <- EOF(gh.t.w ~ lat + lon | date, data = geopotential, n = 1)
+#' aao <- EOF(gh.t.w ~ lat + lon | date, data = geopotential, n = 1,
+#'            B = 100, probs = c(lo2 = 0.1, hig = 0.9))
 #'
 #' # AAO field
 #' library(ggplot2)
@@ -63,7 +69,8 @@
 #' ggplot(aao$right, aes(date, gh.t.w)) +
 #'     geom_line()
 #'
-#' # % of explained variance
+#' # standard deviation, % of explained variance and
+#' # confidence intervals.
 #' aao$sdev
 #'
 #' # 1st eof for each month.
@@ -74,15 +81,17 @@
 #'     facet_wrap(~ month)
 #'
 #' # Alternative interface
-#'
 #' aao2 <- EOF(lon + lat ~ date, value.var = "gh.t.w", data = geopotential)
 #'
+#' @references
+#' Fisher, A., Caffo, B., Schwartz, B., & Zipunnikov, V. (2016). Fast, Exact Bootstrap Principal Component Analysis for p > 1 million. Journal of the American Statistical Association, 111(514), 846–860. http://doi.org/10.1080/01621459.2015.1062383
 #' @family meteorology functions
 #' @export
 #' @import data.table
 #' @import Formula
 #' @importFrom stats as.formula
-EOF <- function(formula, value.var = NULL, data = NULL, n = 1) {
+EOF <- function(formula, value.var = NULL, data = NULL, n = 1, B = 0,
+                probs = c(lower = 0.025, mid = 0.5, upper = 0.975)) {
 
     if (!is.null(value.var)) {
         if (is.null(data)) stop("data must not be NULL if value.var is NULL",
@@ -111,7 +120,6 @@ EOF <- function(formula, value.var = NULL, data = NULL, n = 1) {
 
 
     if (requireNamespace("irlba", quietly = TRUE)) {
-        set.seed(42)
         eof <- irlba::irlba(g$matrix, nv = max(n), nu = max(n), rng = runif)
     } else {
         eof <- svd(g$matrix, nu = max(n), nv = max(n))
@@ -124,19 +132,50 @@ EOF <- function(formula, value.var = NULL, data = NULL, n = 1) {
     right <- cbind(right, as.data.table(g$coldims))
     right <- data.table::melt(right, id.vars = names(g$coldims), variable = "PC",
                               value.name = value.var)
-    right[, PC := factor(PC)]
+    right[, PC := factor(PC, levels = pcomps, ordered = TRUE)]
 
     left <- as.data.table(eof$u[, n])
     colnames(left) <- pcomps
     left <- cbind(left, as.data.table(g$rowdims))
     left <- data.table::melt(left, id.vars = names(g$rowdims), variable = "PC",
                              value.name = value.var)
-    left[, PC := factor(PC)]
+    left[, PC := factor(PC, levels = pcomps, ordered = TRUE)]
 
     v.g  <- norm(g$matrix, type = "F")
     sdev <- data.table(PC = pcomps, sd = eof$d[n])
-    sdev[, PC := factor(PC)]
+    sdev[, PC := factor(PC, levels = pcomps, ordered = TRUE)]
     sdev[, r2 := sd^2/v.g^2]
+
+    if (B > 1) {
+        tall <- dim(g$matrix)[1] > dim(g$matrix)[2]
+        set.seed(42)
+        if (!tall) {
+            names(eof) <- c("d", "v", "u", "iter", "mprod")
+        }
+        # if (tall) {
+            p <- nrow(eof$v)
+            sdevs <- lapply(seq_len(B), function(x) {
+                Prow <- sample(seq_len(p), replace = TRUE)
+                m <- diag(eof$d, nrow = max(n), ncol = max(n))%*%t(eof$v)[, Prow]
+                return(svd(m)$d)
+            })
+        # } else {
+        #     p <- ncol(eof$u)
+        #     sdevs <- lapply(seq_len(B), function(x) {
+        #         Prow <- sample(seq_len(p), replace = TRUE)
+        #         m <- eof$u%*%diag(eof$d, nrow = max(n), ncol = max(n))[Prow, ]
+        #         return(svd(m)$d)
+        #     })
+        # }
+        se <- lapply(data.table::transpose(sdevs), quantile, probs = probs, names = FALSE)
+        se <- data.table::transpose(se)
+        # m <- unlist(lapply(data.table::transpose(sdevs), mean))
+    } else {
+        se <- NA
+    }
+    # sdev[, sd := m]
+    if (is.null(names(probs))) names(probs) <- scales::percent(probs)
+    sdev[, names(probs) := se]
 
     return(list(right = right, left = left, sdev = sdev))
 }