+++ NEWS.md
@@ -17,6 +17,7 @@ communicating the scale of vectors.
 - `stat_contour_fill()` and `stat_contour2()` print a warning when no contours 
 can be made.
 - `geom_contour2()` now has a `gap` parameter to add space for labels.
+- `EOF()` now supports estimation of confidence intervals via bootstrap.
 
 # metR 0.1
 - New function: `GetTopography()`